Fix if statements

+++ user.sh
@@ -12,7 +12,7 @@ fi
 # make symbolic links of dot files
 cd $HOME
 for path_to_file in $(find $DIR -name '.?*' | grep -v '\.git' | grep -v '\.\.' | grep -v '\.gitignore'); do
-  if [ $HOME/$(basename $path_to_file) ]; then
+  if [ -f "$HOME/$(basename $path_to_file)" ]; then
     rm $HOME/$(basename $path_to_file)
   fi
   ln -s $path_to_file $HOME
@@ -20,7 +20,7 @@ done
 cd $HOME
 
 # clone git repositories
-if [ ! $HOME/repo ]; then
+if [ ! -d "$HOME/repo" ]; then
   mkdir $HOME/repo/
 fi
 cd $HOME/repo/
@@ -31,7 +31,9 @@ cd $HOME
 
 # setup vim
 cd $HOME
-rm -rf $HOME/.vim
+if [ -d "$HOME/.vim" ]; then
+  rm -rf $HOME/.vim
+fi
 mkdir -p $HOME/.vim/{autoload,colors,plugged}
 cp $HOME/repo/vim-plug/plug.vim $HOME/.vim/autoload
 cp $HOME/repo/heroku-colorscheme/colors/* $HOME/.vim/colors/
